requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation (Python 3.13 compatible)
pydantic>=1.10.0,<2.0.0
email-validator>=1.3.0
python-dotenv>=1.0.0

# Basic HTTP (no httpx for now to avoid conflicts)
requests>=2.31.0
aiofiles>=23.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development (no pytest to avoid TestClient)
black>=23.0.0
flake8>=6.0.0

# Utilities
typing-extensions>=4.0.0

# Database (SQLite)
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# File handling
aiofiles>=23.0.0

# OpenAI integration
openai>=1.0.0

# Redis integration
redis[async]
fakeredis[async]

# AWS S3 Integration
aioboto3>=15.0.0
boto3>=1.38.0
